Add fmtlib, make it actually work?

+++ bridge/src/bridge.hpp
@@ -3,5 +3,5 @@
 #include <shader_compiler/common/common_types.h>
 
 namespace Bridge {
-    rust::Vec<u8> translate_shader(rust::Vec<u8> shader);
+    rust::Vec<u32> translate_shader(rust::Vec<u8> shader);
 }
